BUGFIX:
PROBLEM: Using Dovecot with MDBOX format of mailbox, if you delete a box with one or more sub-boxes, the dovecot can't remove the dir of parent box.

If the back-end use dovecot with MAILDIR, the DELETE action in mailbox, remove the parent and all children boxes because the structure of files and dir is different of MDBOX ( I dont see the code, but I should look like a "rm -rf" on dir).

MAILDIR:

parent_box
  |_cur
  |_new
  |_tmp
  |_child-box
               |_cur
               |_new
               |_tmp
 
MDBOX:

parent-box
   |_dirs_boxes
   |_child-box
             |_dirs_boxes

To fix in roundcube:
I change the order of DELETE Boxes, first the remove the children box after remove parent box.