New optional HTML/plaintext view icons in #messageheader mean that we need wider CSS margin on the message subject.
Otherwise a long subject line runs behind the extra icons.

This shows up frequently with the threecol plugin installed:

  https://github.com/tofi86/Roundcube-Plugin-Threecol-Layout

(I have submitted a pull request to make that work with 1.0.1)

However you see the same problem with the standard mailbox view
if a message has a very long subject line.