Fix smart column display in rcmail_js_message_list().
This fixes an issue where the Sent folder will show the From address in
the 'fromto' column rather than, as it should, the To address. This
appears to be because the call to
rcmail_message_list_smart_column_name() is made after the folder context
has been overridden, thus getting the wrong folder name and making a
wrong decision about the contents of 'fromto'.

This patch moves the call to rcmail_message_list_smart_column_name() to
be earlier in the rcmail_js_message_list() function and seems to have
fixed the issue for me.

This issue was experienced on the Roundcubemail packages distributed by
the Kolab 3.3 repository (both the recently released 1.1 and the version
prior to that). The fix is also tested against that version.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>