plugins/password/drivers/sql: Add config option for password schemes
Often the password scheme is prepended to the hash. Dovecot is doing
that and the VirtualMailManager (VMM).

Add a boolean option to enable prepending the scheme.

The `doveadm pw` cannot be used, especially if Roundcube is not
installed on the mail server. Actually `password_dovecotpw_with_method`
could be renamed, if that is allowed.