Fix regexp in wash_style() to accept style declaration values with new lines.
Some(?) email clients wrap long lines in HTML messages. With some "bad luck", the wrapping occurs in the middle of a style declaration. This is an example of a style declaration after wrapping:
(...)<div style="padding: 0px
         20px;border:1px solid #000;">(...)
With the "s" modifier in preg_match(), the "." in the regular expression accepts the "\n" and the style declaration remains correct.