Extend date_today config to allow 'Today' label to be removed
The hidden config option date_today allows for the position of the 'Today' label to be changed within the string. This PR extends that functionality to allow the 'Today' label to be removed completely.

I have also cleaned up the format_date function slightly to remove some duplicate code.

Note about backwards compatibility: The date_today config option was removed from config files in 0.7 and since the config files have been entirely rewritten since then I'm not sure this posses much risk for backwards compatibility but it is possible. In this case
```php
$config['date_today'] = '$ ' . $config['time_format'];
```
should be added to the config file.