Update rcube_sieve_vacation.php
Problem encountered where $auto_addr was true and vacation['addresses'] was undefined. This caused a null/empty addresses value. The above check ensures that the addresses value is only used if it is defined and populated.

Adding the !empty check ensures what I believe to be the expected behaviour where $auto_addr is true.