Correct mime_id in case of attachments
This is a temporary fix. When there are attachments, the message is parsed twice. The mime_id becomes 1.2 (first block, second sub-block). Here, mime_id returns always 2, as it is not aware we are in a sub-block. Thus, when there is an attachments, you will not get the "Message decrypted" message, as $this->decryptions is not updated correctly.