jsdeps.sh: Store downloaded files in cache folder, add option to install deps from it
When in git-master you install deps and then want to switch to an older branch
you'll get a conflict. In such a case you can just call `jsdeps.sh cleanup`
and when you're back in master `jsdeps.sh from-cache` which will be much faster
and do not require re-downloading everything.