password plugin: better error message on constraint violation
If the password policy on the ldap server is violated when changing
password, report error as "constraint violation" instead of the
misleading "connection error".

Note: LDAP_CONSTRAINT_VIOLATION can be returned for multiple
reasons – not just password policy violation – but in this case
it's a pretty safe bet it's caused by a ppolicy violation.

Only tested with OpenLDAP (v2.4).