Refactored Dockerfile as suggested in #5874
It's now based on a php:7.1-apache image and pre-configures Roundcube
from the given environment variables to either connect to a MySQL server
or to use a local SQLite database.

Added a README file with some basic information and description of the
most important denv variables. More configuration options are yet to come.

This is a pre-requisite for #5827.