Managesieve allowed hosts (#6292)
* Introduce managesieve_allowed_hosts setting

If managesieve_allowed_hosts has been set, then check that the array contains the current mail_host for the user session. If not, just return. This will cause the managesieve plugin to not appear under the settings menu.

* Include managesieve_allowed_hosts example

An example of the managesieve_allowed_hosts setting is included. By default, it is left commented to preserve existing behaviour.