address_book_type plugins fix
If both LDAP and plugin address books are used, the LDAP one is forced as default (until changed in preferences), even if `address_book_type` is empty.  This patch makes RC check plugins for available address books first, if `address_book_type` is empty. And the plugins address books are now listed before the LDAP ones.