Force HIGHESTMODSEQ use on dovecot
This may fix some messages cache issues. Dovecot does not return
HIGHESTMODSEQ nor NOMODSEQ metadata on SELECT until it is "forced"
to switch to "permanent mod-seq mode". It can be done either by
ENABLE QRESYNC/CONDSTORE or by requesting the metadata explicitely
on SELECT/STATUS. We'll use STATUS for that.