Expire auth cookie after session lifetime
The authentication cookie is regularly renewed while the browser is
active. But when the browser has been inactive for longer than the
session lifetime, it's invalidated. There is no reason to keep the
cookie beyond this point.

On the other hand, the cookie was previously bound to the browser
session. Closing the browser would delete the authentication cookie.
Giving the cookie an expiry date will persist the cookie and can keep
the authentication alive.

Please not though that the session cookie is currently still deleted
when closing the browser. Sessions are not yet persisted but I'm
planning to chane that in a future commit.