Restore compatibility in rcube_addressbook::set_sort_order API
Addressbook plugins used to be either able to either store their
supported column types as a flat list of column names (like
rcube_contacts does), or as an array where the column names are keys and
additional attributes are provided for the columns (e.g. custom
subtypes).

In the current master, only the former type is supported by the
set_sort_order function in the rcube_addressbook class. It used to
supports both.

For the latter type, the original support was a bit broken though, as it
did not allow the sort columns to map to an empty array. This is used in
the carddav addressbook because it only requires custom properties for
some attributes, and sets the others to empty arrays which roundcube
then merges with its standard properties for these columns. This is
fixed by using isset that only checks if the the key exists in coltypes.

I know a plugin can override this function, but this change
unnecessarily breaks backwards compatibility.