Don't write stats info to console.log when in CLI
If the file was created by the web server user, user executing the CLI command
might have no permissions. This leads to an error, and e.g. when installing
a plugin/skin via composer is converted to an ugly exception. We want to prevent that.
Anyway, these stats are most useful for the webmail use not for CLI commands.