Allow to set HTML-part Content-Transfer-Encoding determined by   config "Use MIME for 8bit" setting
Allows to set Content-Transfer-Encoding for HTML part in similiar way as for Text part. (Based on Settings-Writing-Use MIME for 8bit) Previously it was hardcoded quoted-printable.
If it is intentional (bugs in user agents...), please explain. According to my observations nearly everyone uses it and I think I should not be problematic, but I am not sure totally.