Fix a small typo in el_GR/labels.inc
I just happened to notice this tiny typo on a local installation of Roundcube. I just sent a request to join the Transifex organisation to fix this there as well and help with the Greek translation. 

If typo fixes should exclusively be done via Transifex as well, feel free to close this PR instead of merging.