Fix sort_folder_comparator
In case $path2 is longer than $path1, and all folders in $path1 are
equal to the corresponding folder in $path2 (in other words: $path2 is a
subfolder of $path1), sort_folder_comparator currently considers them
equal because it stops comparing when it reaches the end of $path1.

This breaks the ordering done by uasort() above, and breaks code that
relies on the folder list being sorted, notably the folders settings
page.