Initialize Mailvelope earlier (#9243)
This helps fresh users to have a site-specific keyring directly after
the first login (see below). And it hurts no one else because the heavy
Mailvelope lifting works with Promises, which don't block the main code
flow.

If Mailvelope's main keyring is not to be used (configurable in
Roundcubemail's settings), users with a fresh setup previously had only
Mailvelope's main keyring after a login to Roundcubemail. Only after
opening an encrypted message, or going to the compose page, they got a
site-specific keyring. This lead to some users being confused where the
keys went they created after logging in.