Don't render quota if it's unlimited
In case an accounts's quota is zero and the config option
`quota_zero_as_unlimited` is true, then the account has no quota and
should not be shown a meaningless (or even wrong) quota-display.

This introduces a new function to avoid repeating identical code in two
places.

If a template doesn't check for the env-flag it still can render quota
details.