Fix double scrollbar in compose window with big body (#9484, #7760)
If editing a draft, that has a lot of text content already, previously
sometimes a double scrollbar was shown, because the callback function
that calculates the height for the textare was triggered too early.

This change sets the first height calculation behind a 100ms timeout.
(A 0ms timeout might be feasible, too, just to put the callback
execution on the call stack – it also works locally, but I'm not sure
about real world behaviour, so better be safe than sorry.)