Script to test remote access to `temp_dir` and `log_dir`
Those two config options have been set to directories inside the
document root, which is only protected if the HTTP daemon is Apache and
that loads .htaccess files.
So in a lot of situations out there, people's temporary files (which
includes message texts and attachments during while they are edited) and
log files (which might include email adresses) are very public.