Use lax CSP whereever the HTML editor might be used
A "lax" CSP (aka using the config option
`content_security_policy_add_allow_remote`) is required to allow using remote
ressources like image URLs in the HTML editor.

We can't depend this on the intial content being HTML or not because the
user might want to change the editor after loading the page, and then
add remote ressources.