Markmirror: compose in markdown
This adds a markdown editor that sends HTML to the server.

It uses codemirror and some custom code to show a syntax highlighted
textarea and some buttons to help editing
(including a preview).

Drafts get an additional email header that causes the markdown editor to
automatically start if a message composition is
continued that was started using the markdown editor.